backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
psutil>=5.9.0
transformers>=4.37.0
torch>=2.1.0
onnxruntime>=1.17.0
optimum[onnxruntime]>=1.17.0
datasets>=2.16.0
numpy>=1.26.0
pydantic>=2.5.0
scikit-learn>=1.4.0
pytest>=7.4.0
websockets>=12.0
